pyproject.toml
[project]
name = "startup-dued-forge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-mistralai>=1.1.1",
    "praw>=7.8.1",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
]
